backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-adk==1.2.1
google-genai>=1.17.0
httpx==0.28.1
reportlab==4.2.5
python-dotenv==1.0.1
